+++ link/src/main/java/com/meshconnect/link/ui/LinkActivity.kt
@@ -28,17 +28,21 @@ import com.meshconnect.link.databinding.LinkActivityBinding
 import com.meshconnect.link.entity.LinkConfiguration
 import com.meshconnect.link.entity.LinkEvent
 import com.meshconnect.link.entity.LinkPayload
+import com.meshconnect.link.utils.THEME_DARK
 import com.meshconnect.link.utils.alertDialog
 import com.meshconnect.link.utils.createURL
 import com.meshconnect.link.utils.decodeToken
-import com.meshconnect.link.utils.getLinkStyleFromLinkUrl
 import com.meshconnect.link.utils.getOnLoadedScript
 import com.meshconnect.link.utils.getParcelable
+import com.meshconnect.link.utils.getThemeFromUrl
+import com.meshconnect.link.utils.getThemeName
 import com.meshconnect.link.utils.intent
-import com.meshconnect.link.utils.isSystemDarkTheme
+import com.meshconnect.link.utils.isSystemThemeDark
 import com.meshconnect.link.utils.isUrlWhitelisted
 import com.meshconnect.link.utils.observeEvent
 import com.meshconnect.link.utils.openTrueAuth
+import com.meshconnect.link.utils.resolveLanguage
+import com.meshconnect.link.utils.resolveTheme
 import com.meshconnect.link.utils.showToast
 import com.meshconnect.link.utils.viewBinding
 import com.meshconnect.link.utils.viewModel
@@ -52,6 +56,8 @@ internal class LinkActivity : AppCompatActivity() {
         private const val DISABLE_WHITELIST = "disable_whitelist"
         private const val DATA = "data"
         private const val LANGUAGE = "language"
+        private const val FIAT_CURRENCY = "fiat_currency"
+        private const val THEME = "theme"
         private const val MAX_TOAST_MSG_LENGTH = 38
         private const val CBW_HOST = "wallet.coinbase.com"
         private const val CBW_PACKAGE_NAME = "org.toshi"
@@ -65,6 +71,8 @@ internal class LinkActivity : AppCompatActivity() {
                     .putExtra(TOKEN, config.token)
                     .putExtra(DISABLE_WHITELIST, config.disableDomainWhiteList)
                     .putExtra(LANGUAGE, config.language)
+                    .putExtra(FIAT_CURRENCY, config.displayFiatCurrency)
+                    .putExtra(THEME, getThemeName(config.theme))
 
             val accessTokens = config.accessTokens
 
@@ -88,11 +96,14 @@ internal class LinkActivity : AppCompatActivity() {
 
         runCatching {
             val link = decodeToken(intent.getStringExtra(TOKEN))
-            createURL(link, mapOf("lng" to intent.getStringExtra(LANGUAGE)))
+            val language = resolveLanguage(intent.getStringExtra(LANGUAGE))
+            val fiatCurrency = intent.getStringExtra(FIAT_CURRENCY)
+            val theme = resolveTheme(intent.getStringExtra(THEME), this::isSystemThemeDark)
+            createURL(link, mapOf("lng" to language, "fiatCur" to fiatCurrency, "th" to theme))
         }.onSuccess { url ->
             setContentView(binding.root)
 
-            applyTheme(url.toString())
+            applyThemeFromUrl(url.toString())
 
             binding.back.setOnClickListener { onBack() }
             binding.close.setOnClickListener { showCloseDialog() }
@@ -110,14 +121,10 @@ internal class LinkActivity : AppCompatActivity() {
         }
     }
 
-    private fun applyTheme(url: String) {
-        val linkStyle = getLinkStyleFromLinkUrl(url)
-        val isDarkTheme =
-            when (linkStyle?.th) {
-                "dark" -> true
-                "system" -> isSystemDarkTheme()
-                else -> false
-            }
+    private fun applyThemeFromUrl(url: String) {
+        val theme = getThemeFromUrl(url)
+        val isDarkTheme = resolveTheme(theme, this::isSystemThemeDark) == THEME_DARK
+
         val topColor = getColor(if (isDarkTheme) R.color.coldGray70 else R.color.coldGray0)
         val bottomColor = getColor(if (isDarkTheme) R.color.gray80 else R.color.gray0)